Unify a single Election page
