+++ src/App.tsx
@@ -3,21 +3,16 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 import 'normalize.css';
 
 import Layout from './components/Layout';
-import InputTrackerPage from './pages/InputTrackerPage';
-import ResultsPage from './pages/ResultsPage';
 import NotFoundPage from './pages/NotFoundPage';
 
 import './App.css';
+import ElectionPage from './pages/ElectionPage';
 
 function App() {
     return (
         <Layout>
             <Switch>
-                <Route path="/" exact>
-                    <Redirect to="/confirm-ballot" />
-                </Route>
-                <Route path="/confirm-ballot" component={InputTrackerPage} />
-                <Route path="/results" component={ResultsPage} />
+                <Route path="/" exact component={ElectionPage} />
                 <Route component={NotFoundPage} />
             </Switch>
         </Layout>
